requirements.txt
pandas
plotly.graph_objects
numpy
matplotlib